==> code/00_library.R <==

################          LIBRARIES USED         ################


library(googlePolylines)#!
library(paletteer)
library(leaflet)
library(ecmwfr)
library(htmltools)

library(ncdf4)
library(dplyr)
library(sp)
library(raster)
library(sf)
library(maps)
library(mapdata) #!!!
library(maptools)#!!!
library(rgdal)

library(gdalUtilities)#!!!
####library(gpclib) #!!!
library(spatstat)#!!!
####library(RGtk2)
library(MODIStsp)
####library(MODIS)
library(ggplot2)
library(ggmap)
library(lubridate)
library(rgeos)
library(reshape2)#!!!
library(dplyr)
library(lmtest)#!!!
library(tm)#!!!
library(csv)#!!!
library(csvread)#!!!
library(R.utils)#!!!
library(lattice)
library(caret)#!!!
#### library(RStoolbox)#!!!
library(RColorBrewer)
library(openair)
library(gridExtra)
library(printr)#!!!
library(corrplot)#!!!
library(randomForest)#!!!

####library(xlsx)#!!!
library(tdr)#!!!
library(tidyverse)#!!!
library(ggExtra)#!!!
#library(ggpp)#!!!
library(ggmap)
library(ggpubr)#!!!
library(readr)
library(tidyverse)#!!!
library(httr)
library(jsonlite)
library(ggmap)
library(gepaf)#!!!
library(dplyr)


library(lubridate)
library(foreach)
library(googlePolylines)
library(paletteer)
library(leaflet)
library(terra)
library(stars)

library(htmltools)

#Configurar API
https://console.cloud.google.com/projectselector2/google/maps-apis/credentials?pli=1
#https://console.cloud.google.com/apis/library?supportedpurview=project


## dar de baja a la cuetna

https://www.google.com/search?q=dar+de+baja+cuenta+google&oq=dar+de+baja+cuenta+goo&aqs=chrome.0.0i512j69i57j0i512l3j0i22i30l5.591737j0j7&sourceid=chrome&ie=UTF-8
